Changed viewer page to main/home page
